requirements.txt
streamlit
pandas
openpyxl
google-genai
xlrd
pypdf
python-docx